# internal/handler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Constellation Control Plane API",
    "version": "0.1.0",
    "description": "Phase 1 control-plane endpoints. Astronomer integration surfaces /api/v1/security/* alongside the standalone /api/v1/* routes."
  },
  "servers": [
    {
      "url": "http://localhost:18080"
    },
    {
      "url": "https://constellation.example.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error envelope returned by API handlers.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Get groups"
      },
      "post": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Create groups"
      }
    },
    "/api/v1/groups/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Delete groups"
      },
      "put": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Replace groups"
      }
    },
...
    "/api/v1/groups:export": {
      "get": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Get groups:export"
      }
    },
    "/api/v1/groups:import": {
      "post": {
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Create groups:import"
      }
    },
    "/api/v1/groups:promote": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dimension",
                  "from"
                ],
                "properties": {
                  "dimension": {
                    "type": "string",
                    "enum": [
                      "policy",
                      "profile"
                    ]
                  },
                  "from": {
                    "type": "string",
                    "enum": [
                      "discover",
                      "monitor",
                      "protect"
                    ]
                  },
                  "to": {
                    "type": "string",
                    "enum": [
                      "discover",
                      "monitor",
                      "protect"
                    ],
                    "description": "Optional explicit target mode. When omitted, discover advances to monitor and monitor advances to protect."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk group mode change summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "dimension",
                    "from",
                    "to",
                    "changed",
                    "promoted"
                  ],
                  "properties": {
                    "dimension": {
                      "type": "string"
                    },
                    "from": {
                      "type": "string"
                    },
                    "to": {
                      "type": "string"
                    },
                    "changed": {
                      "type": "integer"
                    },
                    "promoted": {
                      "type": "integer",
                      "description": "Backward-compatible alias for changed."
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "Bulk change group policy/profile modes"
      }
    },
...
  }
}